backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
faiss-cpu
pillow
transformers
sentence-transformers
